pyproject.toml
[project]
name = "lovable-bridge-mcp"
version = "0.1.0"
description = "MCP Server for importing Lovable projects into Databricks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "mcp>=0.9.0",
    "pydantic>=2.6.0",
    "databricks-sdk>=0.23.0",
    "sqlmodel>=0.0.16",
    "alembic>=1.13.0",
    "jinja2>=3.1.3",
    "tree-sitter>=0.21.0",
    "tree-sitter-typescript>=0.21.0",
    "sqlparse>=0.4.4",
    "httpx>=0.26.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.41",
    "orjson>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "black>=24.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
